refactor to avoid duplicated Kuberenetes outputs
